fix(workflowengine):update stale dependencies
