+++ core/src/lib/workflow-event.ts
@@ -1,0 +1,13 @@
+export enum WorkflowEventType {
+    beginWorkflow,
+    beginStep,
+    endStep,
+    endWorkflow,
+}
+
+export interface WorkflowEvent {
+    eventType: WorkflowEventType;
+    stepLabel?: string;
+    stepIndex?: number;
+    workflowName: string;
+}
